:recycle: Refactor CDC to split avro and protobuf function
