+++ .gitignore
@@ -162,3 +162,4 @@ cython_debug/
 # Spark folders
 spark-warehouse/
 metastore_db/
+*.desc
